move tests to UserInformationControllerTest, remove UserInformationControllerIT, remove hamcrest
