+++ web-common/pom.xml
@@ -136,13 +136,6 @@
           <artifactId>jakarta.annotation-api</artifactId>
           <version>${jakarta.annotation.version}</version>
       </dependency>
-      <!-- Used by spring-test for mvc assertions-->
-      <dependency>
-          <groupId>org.hamcrest</groupId>
-          <artifactId>hamcrest</artifactId>
-          <version>2.2</version>
-          <scope>test</scope>
-      </dependency>
         
       <!-- DCR -->
       <dependency>
